Move path management from e2e_node to common test/utils directory
enables reuse of these methods for other e2e tests
